Rethink percentage bars
for nested directories we end up 'using' all the grey scales early on
and we don't get to see them in the output.

The fix provided here is that if the percent bar generated is the same
as the one that came before it then we don't need to consider a grey
scale color as 'used'.